moved filter-overlay to filter module
