+++ src/app/core/filter/filter.module.ts
@@ -6,10 +6,11 @@ import { ListFilterComponent } from "./list-filter/list-filter.component";
 import { MatFormFieldModule } from "@angular/material/form-field";
 import { MatSelectModule } from "@angular/material/select";
 import { CommonComponentsModule } from "../common-components/common-components.module";
+import { FilterOverlayComponent } from "./filter-overlay/filter-overlay.component";
 
 @NgModule({
-  declarations: [FilterComponent, ListFilterComponent],
-  exports: [FilterComponent, ListFilterComponent, ListFilterComponent],
+  declarations: [FilterComponent, ListFilterComponent, FilterOverlayComponent],
+  exports: [FilterComponent],
   imports: [
     CommonModule,
     Angulartics2Module,
